\name{AbsolutelyStacked-package}
\alias{AbsolutelyStacked-package}
\alias{AbsolutelyStacked}
\docType{package}
\title{
\packageTitle{AbsolutelyStacked}
}
\description{
\packageDescription{AbsolutelyStacked}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AbsolutelyStacked}
\packageIndices{AbsolutelyStacked}
This is a placeholder for documentation to come
}
\author{
\packageAuthor{AbsolutelyStacked}

Maintainer: \packageMaintainer{AbsolutelyStacked}
}
\references{
No references needed here (yet)
}

\keyword{ package }

\examples{
library(AbsolutelyStacked)
print("We should have some example")
}
