% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{fix_inf}
\alias{fix_inf}
\title{Operation that makes 0 times infinity = 0, useful for modelling purposes.}
\usage{
fix_inf(a, b)
}
\arguments{
\item{a}{a real number possibly equivalent to 0.}

\item{b}{an integer, numeric, vector, matrix, or other data type that can be scaled.}
}
\value{
The object b scaled by a.
}
\description{
Operation that makes 0 times infinity = 0, useful for modelling purposes.
}
\details{
This function ensures that when multiplying something going to infinity by something equivalent to 0, 0 wins out (the entire expression becomes 0).
}
\examples{
\donttest{

fix_inf(1/0,0)
}
}
