% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{ifelse2}
\alias{ifelse2}
\title{A straightforward non-vectorized way of extending the flexibility of ifelse}
\usage{
ifelse2(cond, x, y)
}
\arguments{
\item{cond}{condition to evaluate boolean for.}

\item{x}{number of variables randomly sampled as candidates at each split.}

\item{y}{should sampling of cases be done with or without replacement?}
}
\value{
Object of arbitrary class.
}
\description{
A straightforward non-vectorized way of extending the flexibility of ifelse
}
\details{
This increases the flexibility of ifelse when potential inputs and outputs are of different form and/or length.
}
\examples{
\donttest{

  ifelse2(2 > 3, "2 > 3", "3 > 2")
}
}
