% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{kronk}
\alias{kronk}
\title{kronecker product a list of matrices into one block-diagonal}
\usage{
kronk(matlist)
}
\arguments{
\item{matlist}{a list of matrices to Kronecker Product.}
}
\value{
A matrix, the inverse (or Moore-Penrose generalized inverse of) the matrix inputted.
}
\description{
kronecker product a list of matrices into one block-diagonal
}
\details{
This function is helpful for creating large, diagonal matrices given some list of individual, possibly non-conformable matrices.
}
\examples{
\donttest{

kronk(lapply(1:2,function(j)(sapply(1:10,function(i)rnorm(10)))))
}
}
