% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dummy_Z_and_sigmalist.R
\name{make_dummy_Z_and_sigmalist}
\alias{make_dummy_Z_and_sigmalist}
\title{Create dummy random-effects components for ensuring compatibility with glmer_constrained() when only fixed-effects are desired.}
\usage{
make_dummy_Z_and_sigmalist(X)
}
\arguments{
\item{X}{design matrix of fixed effects for the model, or a list for 3+ categorical outcomes}
}
\value{
a list of elements Z, sigma_list, sigmas, and initial random effect values that when inputted to the eponymous arguments of glmer_constrained() function, will not affect estimation of fixed effects.
}
\description{
Create dummy random-effects components for ensuring compatibility with glmer_constrained() when only fixed-effects are desired.
}
\examples{
\donttest{

   ## ordinary vector-valued outcomes
 make_dummy_Z_and_sigmalist(sapply(1:100,function(i)rnorm(100)))

   ## matrix-valued outcomes, i.e., categorical regression
 make_dummy_Z_and_sigmalist(lapply(1:100,function(i)cbind(rnorm(100),rnorm(100))))

}
}
