% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_loglikelihood_MLMetric.R
\name{make_loglikelihood_MLMetric}
\alias{make_loglikelihood_MLMetric}
\title{Convert the log-likelkihood for common REF models into MLMetrics()}
\usage{
make_loglikelihood_MLMetric(link_function, tau, double)
}
\arguments{
\item{link_function}{the link function corresponding to the log-likelihood desired.}

\item{tau}{the fixed exponential dispersion parameter for computing the log-likelihood (i.e., this would be sigma squared for a normal distribution).}

\item{double}{whether to double the log-likelihood when computing the metric, which suggests a Wilks-like statistic.}
}
\value{
an MLMetric that can be used with other MachineShop compatible functions.
}
\description{
Convert the log-likelkihood for common REF models into MLMetrics()
}
\details{
Make MachineShop-compatible MLMetrics reflecting certain log-likelihoods for predicted and observed responses.
}
\examples{
\donttest{

 make_loglikelihood_MLMetric(link_function = 'linear', tau = 1, double = F)
}
}
\seealso{
\code{\link[MLMetric]{MLMetric}}
}
