% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_glmerStacked.R
\name{predict_glmerStacked}
\alias{predict_glmerStacked}
\title{Making predictions on new data with a stacked model, which requires a new Z matrix if random effects are desired.}
\usage{
predict_glmerStacked(fit, newdata, newZ = NULL, use_randint = F, type = "prob")
}
\arguments{
\item{fit}{a fit object returned by running the SuperModel() function with make_glmerStackedModel() as the super learner.}

\item{newdata}{a dataframe of new data for making predictions on}

\item{newZ}{a matrix coding the random effects design matrix for the new data}

\item{use_randint}{T or F, whether or not to include random effects for predicting on future data}

\item{type}{the MachineShop response-type to output for predictions}
}
\value{
out-of-sample predictions made on new data
}
\description{
Making predictions on new data with a stacked model, which requires a new Z matrix if random effects are desired.
}
\examples{
\donttest{

 predict_glmerStacked(fit,newdata = newdata, newZ = diag(1,10))
}
}
