% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{quick_inv}
\alias{quick_inv}
\title{quick way to invert (or alternatively use ginv to find the generalized inverse of) a non-negative definite symmetric matrix utilizing the singular value decomposition}
\usage{
quick_inv(mat)
}
\arguments{
\item{mat}{a symmetric square matrix to be inverted (or find generalized inverse of).}
}
\value{
A matrix, the inverse (or Moore-Penrose generalized inverse of) the matrix inputted.
}
\description{
quick way to invert (or alternatively use ginv to find the generalized inverse of) a non-negative definite symmetric matrix utilizing the singular value decomposition
}
\details{
This function ensures that when multiplying something going to infinity by something equivalent to 0, 0 wins out (the entire expression becomes 0).
}
\examples{
\donttest{

quick_inv(sapply(1:100,function(i)rnorm(100)))
}
}
