% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepAIC_stacked.R
\name{stepAIC_stacked}
\alias{stepAIC_stacked}
\title{Perform Backwards Variable Selection via AIC when stacking.}
\usage{
stepAIC_stacked(
  vars,
  cv_data,
  Zmodel,
  link_func,
  est_tau_model = F,
  t_init,
  s_init,
  b_init,
  u_init,
  list_of_sigmas,
  qp = F,
  min_num_learners,
  use_fixed_weights = F,
  preset_weights = NULL,
  tr = T
)
}
\arguments{
\item{vars}{a character vector of names of variables (i.e. base-learners) to perform selection on.}

\item{cv_data}{a dataframe containing the cross-validated out-of-fold predictions of base learners.}

\item{Zmodel}{the random effects design matrix.}

\item{link_func}{the link function specified.}

\item{est_tau_model}{T or F, is the exponential dispersion parameter to be estimated?}

\item{t_init}{real number greater than 0, the initial (or fixed, if not being estimated) value of the exponential dispersion parameter.}

\item{s_init}{vector of real numbers greater than 0, the initial (or fixed, if not being estimated) values of the random-effect variance parameters.}

\item{b_init}{vector of real numbers, the initial values of weights for estimating unconstrained MLEs.}

\item{u_init}{vector of real numbers, the initial values of random effects.}

\item{list_of_sigmas}{the sigma list, containing the names and corresponding column-indices of the Z matrix corresponding to each random effect variance component.}

\item{qp}{T or F, should sequential quadratic programming be used over a more general non-linear optimizer?}

\item{min_num_learners}{integer > 1, minimum number of base learners to include in stacked model before halting the stepwise selection procedure.}

\item{use_fixed_weights}{T or F, whether stacked model should be fit or fixed weights assigned.}

\item{preset_weights}{If use_fixed_weights == T,then this must be a vector of weights with length exactly equal to the number of base learners considered.}

\item{tr}{T or F, should the program print out stepwise base-learner selection output?}
}
\value{
A MachineShop custom MLModel that can be passed into the SuperModel function as the meta (aka super) learner.
}
\description{
Perform Backwards Variable Selection via AIC when stacking.
}
\details{
Make an MLModel to be inserted into the SuperModel function as the meta (aka super) learner. This allows one to stack ML models for a variety of different outcomes based on the GLM framework, and allows the user to perform backwards variable selection via AIC to choose an optimal combination of base learners.
}
\examples{
\donttest{

 print("Not yet finished")
}
}
\seealso{
\code{\link[MLModel]{MLModel}}, \code{\link[SuperModel]{SuperModel}}
}
